chann: document unbounded channel memory model
The TestNonblockSelectRace test was taken from the runtime package and
were originally for testing buffered channel memory model semantics.
However, the relevant semantics is not directly applicable for the
implementation of the unbounded channel.

This change documents the correct memory model semantics of an unbounded
channel and fixes the flaky test TestNonblockSelectRace.

Fixes #1

Signed-off-by: Changkun Ou <hi@changkun.de>